:recycle: Refactor checked tags feature
